per field double buffering for self channels
This fixes a major problem of only being able to send a value
to only the next instantiation of self. That is, if self ran
before reading the sent value, that value is lost. That was
due to keeping the double-buffer pair switching per channel
and switching it on each task execution. Now the buffer
index is kept per field and switched on OUT.