Revert "make TASK_REF macro not take address"
This reverts commit c31a5d6ffc2cfcd54b47747f6edf34c524d514ea.

Turned out that we still need to define a pointer that
cointains the address of this symbol, and pass that pointer
to CHAN_OUT which will then take its address. To create
the pointer we would need &TASK_REF(), so just factor in the &.
We can't pass the task symbol directly to CHAN_OUT because we need an
address of a pointer to that task object.