clear dirty field count on task restarts
Yes, we don't want to run the prologue on task restarts (aka.  reboots),
we want to run it only on task transtitions. However, on task restarts,
we must clear any state that the incomplete task might have modified
that could be visible to the restarted task instance. The dirty field
counter is such state.

This fixes fatal bug: infinite loop due to overflow of the dirty field
list followed by out of bounds access in prologue.